+++ R/el_lm.R
@@ -41,7 +41,7 @@
 #'   \deqn{H_0: \theta_1 = \theta_2 = \cdots = \theta_{p-1} = 0,}
 #'   and significance tests for each parameter with
 #'   \deqn{H_{0j}: \theta_j = 0,\ j = 0, \dots, p-1.}
-#'   The test results are returned as `optim` and `parTests`, respectively.
+#'   The test results are returned as `optim` and `sigTests`, respectively.
 #' @return An object of class of \linkS4class{LM}.
 #' @references Owen A (1991).
 #'   “Empirical Likelihood for Linear Models.”
@@ -168,7 +168,7 @@ el_lm <- function(formula,
     )
   }
   new("LM",
-    parTests = lapply(out$par_tests, setNames, pnames), call = cl, terms = mt,
+    sigTests = lapply(out$sig_tests, setNames, pnames), call = cl, terms = mt,
     misc = list(
       intercept = intercept,
       xlevels = .getXlevels(mt, mf),