+++ R/getSigTests-methods.R
@@ -1,0 +1,4 @@
+#' @rdname getSigTests
+setMethod("getSigTests", "LM", function(object) {
+  object@sigTests
+})