+++ src/test_LM.cpp
@@ -61,13 +61,14 @@ Rcpp::List test_LM(const Eigen::Map<Eigen::MatrixXd> &x,
     logl = el.loglik();
   }
 
-  // parameter tests
+  // significance tests
   std::vector<double> chisq_val(p);
+  std::vector<int> par_iter(p);
   std::vector<bool> par_conv(p);
   const double test_th = set_threshold(1, th);
-#ifdef _OPENMP
-#pragma omp parallel for num_threads(nthreads)
-#endif
+  #ifdef _OPENMP
+  #pragma omp parallel for num_threads(nthreads)
+  #endif
   for (int i = 0; i < p; ++i)
   {
     Eigen::MatrixXd lhs = Eigen::MatrixXd::Zero(1, p);
@@ -75,12 +76,14 @@ Rcpp::List test_LM(const Eigen::Map<Eigen::MatrixXd> &x,
     const CEL par_test("lm", par0, x, lhs, Eigen::VectorXd::Zero(1), maxit,
                        maxit_l, tol, tol_l, gamma, test_th, w);
     chisq_val[i] = 2.0 * par_test.nllr;
+    par_iter[i] = par_test.iter;
     par_conv[i] = par_test.conv;
   }
 
   Rcpp::List result = Rcpp::List::create(
-      Rcpp::Named("par_tests") = Rcpp::List::create(
+      Rcpp::Named("sig_tests") = Rcpp::List::create(
           Rcpp::Named("statistic") = chisq_val,
+          Rcpp::Named("iterations") = par_iter,
           Rcpp::Named("convergence") = par_conv),
       Rcpp::Named("optim") = Rcpp::List::create(
           Rcpp::Named("par") = par,